+++ NEWS.md
@@ -3,6 +3,7 @@
 # vcfR 1.2.0.9000
 Released on CRAN 2016-XX-XX.
 
+* Created `masplit()` to parse matrices contains delimited strings.
 * Created `ordisample()` to ordinate sample information.
 * `extract.gt()` can now use the ID column from the fix region for rownames.
 * Created `INFO2df()` and `metaINFO2df()`.